code-examples/transaction-monitoring/backend-python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sock==0.7.0
python-dotenv==1.0.1 